Move Cluster component to basics
